US 4:Add some snapshots from decision tree algorithm runs.
Add code to compute the number of features.
This is basically the length of the first element of the features_train array.
SkLearn's SelectPerecentile method provides the option to select the percentage of features.
By default 10% of the features are selected.
Search for this method in the tools /email_preprocess.py file