US:First commit on the enron corpus of data.
Find the length of the python dictionary to know the data size.
Katie cleaned up the data for us and converted it into a dictionary.
Note that the line space issue for python has a workaround as discussed in the forum.
You need to run the docs2unix.py which generates the correct data set python file.